Adapt core.yaml file to 1.2

+++ python-scripts/util.py
@@ -62,10 +62,11 @@ def dowait(what, secs, logFile, files):
 
 def removeIntermediateCerts(intermediatecerts_dir):
     print('Delete intermediate certs in ' + intermediatecerts_dir, flush=True)
-    for file in os.listdir(intermediatecerts_dir):
-        file_path = os.path.join(intermediatecerts_dir, file)
-        if os.path.isfile(file_path):
-            os.remove(file_path)
+    if os.path.exists(intermediatecerts_dir):
+        for file in os.listdir(intermediatecerts_dir):
+            file_path = os.path.join(intermediatecerts_dir, file)
+            if os.path.isfile(file_path):
+                os.remove(file_path)
 
 
 def completeMSPSetup(org_msp_dir):
